Added initialization support to declare_global().
declare_global() now supports the specification of an initial value for
such global. This is not merely syntactic sugar, though it is that as
well, reducing redundancy in global declarations throughout the codebase.
Such support permits declaration of global constants, which, being
constant, require initialization at declaration time.

For typeability, replaced all here-document delimiters with a
reasonably shorter string.