Corrected is_stdin_nonempty().
The prior implementation of such function was essentially fetid poo in
zsh form; the current implementation is mildly less objectionable,
leveraging kernel function select() to test the non-emptiness of
standard input in a non-blocking (and hopefully portable) manner. And
*OH BOY* is the new implementation knackered to all hell. I mean, just
awk through that nonsensical gibberish. (The truth be half-told, we
merely half-grok it ourselves...)