Continued overhauling exception handling.
Sadly, such handling remains fervently broken. This is hardly unexpected. On the
contrary, the broken grandure is an enduring testament to the depth of my zeal
for working full time for no money for no happy ending other than grim
determination. I will see this through, if it kills me.