Defined display server testers statically.
@{*-test/gui/*-display_server} testers (e.g., :is_display_server_x()) are now
defined in a similar manner to those of @{*-test/*-isa} and @{*-test/*-os},
which is to say to unconditionally report either success or failure. Added
detection support for Surface Flinger, the curiously titled display server for
Google Android. Added placeholder parcel @{progress}, to be subsequently
implemented as a dynamic terminal progress bar displayer.