zeshy startup halts on fatal error.
Prior to this commit, zeshy startup failed to halt on fatal error, typically
inducing an unreadable mass of stack traces when merely one would have
sufficed. While this commit corrects such issue, zeshy startup still appears to
fail to halt on thrown exceptions -- which is still somewhat noxious, but not
quite as deplorable.