Zsh >= 5.0.5 compatibility.
An as-yet-undetermined of zsh >= 5.0.5 significantly broke backward
compatibility with respect to parsing of global aliases embedded in process
substitutions. Specifically, global aliases following "}" delimiters in
such substitutions are now erroneously parsed as syntactic errors rather
expanded as expected. Until this is resolved, we defer to the manual
approach and avoid trailing global aliases in process substitutions.

    Tumbling roseate sickle
     cycles inchoate prose
    Fumbled in a die-bold-emboldened
    Obscuration of
    Chiaroscuro arisen.