Zsh >= 5.0.5 compatibility x 2.
Unfortunately, an as-yet-undetermined of zsh >= 5.0.5 significantly
broke backward compatibility in a variety of respects -- not merely
the parsing of global aliases embedded in process substitutions
addressed by the prior commit.

It would additionally appear that TRAPZERR handling, which zeshy heavily
depends upon, has also been subtly broken. Specifically, zsh now triggers
the TRAPZERR handler if *ANY* command in a sublist contained in the "else"
branch of an "if" conditional that is itself the last command in a
function reports non-zero exit status regardless of whether that command
is the last command in that sublist or not. This common case appears
sufficiently frequently throughout the zeshy codebase that the zeshy
codebase has now effectively become horribly broken -- and we have no
reasonable means of addressing this, short of simply prohibiting use of
newer versions of zsh.

In short:

    So
    Much
    Rage,
    Right
    Now.